Use PostingsForMatchersProvider in headIndexReader
There's no risk in doing this as we're not introducing anything new:
callers could hold the postings previously while head was truncated.
It's just that series won't be found when referenced with non-existing
postings.

In any case, this uses other methods implemented on headIndexReader,
that should have already implemented the needed protection.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>