model/labels: Make Labels type a struct
Implementation is the same underneath: this is a first step to flush out
all the parts of Prometheus that rely on being able to access Labels as
a slice.

To fix YAML marshalling, since Labels now has no exported fields, it
needs to implement `IsZero()` otherwise 'omitempty' fields are always
omitted.

The change to New() was requested by lint; not sure why it didn't
complain earlier.

Signed-off-by: Bryan Boreham <bjboreham@gmail.com>