Extend memseries to hold OOO Chunks data
In an effort to extend prometheus support for OOO samples one of the
first steps is to extend the memSeries struct to hold information about
the chunks containing out of order samples.

Also left a few TODO notes in the db.go file. Some of the next steps
will consist on creating a NewOOORangeHead method and then extend the db
file to compact head ooo data.