Start implementing OOOHead
Very much like RangeHead we're now implementing an equivalent for the
OOO samples in the Head.

OOOHead implements BlockReader interface which has methods such as
Index(), Chunks(), Tombstones(), ... and that lead to create
oooHeadIndexReader to be able to provide a functional implementation for
the Index() method.

oooHeadIndexReader implements IndexReader and now we're finally using
some of the fields we added to the head to keep OOO samples. Leaving
this PR here because I want to carefully review the changes so far.

The same would have to be done to satisfy Chunks but that will come in
future commits.