Add a head index reader reference to oooHeadIndexReader
While implementing the IndexReader interface for oooHeadIndexReader, we
noticed we only need to change a few methods so duplicating the
remaining ones felt a bit of an overhead.

To avoid this duplication of code we've decided to add a reference to
the headIndexReader to the struct so we can leverage on it the calls for
those methods that share the same implementation.