Actually delegate most of the methods to headIndexReader
We've decided to go with the idea that the OOOHead doesn't need to
filter out its labels and postings responses to only series in the head
with ooo samples.

And with the same reasoning the RangeHead doesn't need to do it for only
series without ooo samples.

Instead we take the easiest path forward and OOOHead delegates the
implementation of these methods to the RangeHead.