Initial implementation of OOOHeadIndexReader Series()
As stated in https://github.com/grafana/mimir-prometheus/issues/147#issuecomment-1048064175
we needed the Series() method of OOOHeadIndexReader to handle
overlapping OOO chunks so in this commit I'm trying to address the
proposal suggested.

Series() will collect available ooo head and memory mapped chunks, order
them based on mintime and then filter out those that are overlapping.

In the end Series() will return the chunk Metas of the first overlapping
chunks as well as the metas of chunks that do not overlap.