Initial ChunkReader implementation for OOOHeadChunkReader
We've used the existing implementation of ChunkReader from
headChunkReader with a few minor changes.

Instead of using the chunk() method we've created oooChunk() because
the logic will differ just how it did it in the BlockReader
implementation. oooChunk() does not only need to return the chunk
matching the given id but actually a merged chunk containing that one
and any possible overlapping chunks.

Because of this we need to have this new method and implement the logic
there.

This commit does not actually implement the correct logic. It does set
the playground to do it and a bunch of TODOs that need answer before
moving forward.