oooMergedChunk now returns a mergedOOOChunk
New type created called mergedOOOChunks which holds a list of chunk
metas and implements the Chunk interface.

oooMergedChunk now finds all the overlapping chunks with the given chunk
meta and composes a mergedOOOChunk object with them.

Then oooMergedChunk implements all the methods in Chunk interface
including Iterator which relies on the existing implementation of
storage.NewChainSampleIterator to compose a high level iterator of
sorted samples.

Missing unit tests which will come in a future commit.