Add overlapping chunks in decreasing order test
We thought about a few more cases that we think the Series() method
should be able to support.

This particular commit brings one where 4 chunks overlap in pairs in
decreasing order.

I think we need to review the assumption we made about leaving the head chunk
max time set to maxint again because if the head chunk contains the oldest
samples in the interval then our algorithm fails.

In this commit I'm setting the maxtime to the headchunk's existing maxtime
instead.