fix which markers the tests use
should be the last interval to overlap, not the last one defined.
so our choice is to either compute it automatically in the test runner,
but that would replicate a significant amount of the code under test,
undermining the the test, or specify it manually in the tests again,
which we now do.