Exclude ooo head samples that were appended after
Updates the Chunk() implementation to leave head chunk samples out if
they are out of the last known OOOLastMinT and OOOLastMaxT markers that
were passed in the meta.

This comes from an scenario we imagined in
https://github.com/grafana/mimir-prometheus/issues/147. It's actually
the 5th one where the head could get new samples in between Series() and
Chunk() call.