clarify OOO metrics and comments
In headAppender.Commit(), we always report any delta into the OOO
histogram, regardless of whether the sample was accepted or rejected,
and regardless of whether OOO is enabled or not.

In headAppender.Append(), we only reported the delta for OOO samples and
when OOO support was disabled.

The comprehensive behavior seems more useful, so switch to that.