labels: shrink by making internals a single string
Format is:

    [len][name0][len][value0][len][name1][len][value1]...

The lengths are varint encoded so usually a single byte.

The previous `[]string` had 24 bytes of overhead for the slice and 16
for each label name and value; this one has 16 bytes overhead plus 1
for each name and value.

In `Overwrite` and `Hash` we use an unsafe conversion from string to
byte slice. `Overwrite` is explicitly unsafe, but for `Hash` this is
a pure performance hack.

Signed-off-by: Bryan Boreham <bjboreham@gmail.com>