Hackathon: raw_increase
Introduces a new function `raw_increase(matrix, scalar millis)` that
calculates the raw incrase ("raw" as opposed to the extrapolated
increase that `increase()` and `rate()` calculate) in the window
provided by the scalar millis param. It can make use of a bigger matrix
range if there are not enough values in the window, to find the previous
value.

It accounts for the counter resets.

If there are at least 2 samples in the window, it doesn't look back past
the window.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>