OTLP: Increase valid created timestamp threshold to 5 minutes
Increasing valid threshold for inserting zero's at StartTimeUnixNano on OTLP
Write requests.

The reason why we want to do this is that Grafana's $__rate_interval is
typically 4 times the scrape interval so assuming an average DPM of 1 we could
have the zero sample 4 minutes before the first sample.

This is a small change that might help with usecases where samples are slightly
delayed.

Also we've made this threshold configurable now. It defaults to 5 minutes but
it can be overiden for special cases.

Signed-off-by: Jesus Vazquez <jesus.vazquez@grafana.com>