Implement index.MaybeRecyclePostings
This method would call an optional Recycle() method on the postings,
that would reuse the reference by putting it into a pool.

The heuristic I've followed is that it should be called by the method
that performs the Next() check, as it's that method who knows that
postings can't be used anymore.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>