fix(tsdb): Clear reset hint when switching between ino and ooo head
Clear the native histogram counter reset hint when switching between the
in-order and out-of-order head.

Given the samples written in this order:
t=1, v=40
t=4, v=30  (reset in in-order)
t=2, v=40
t=3, v=10  (reset in out-of-order)

When we read them back, this is the order:
t=1, v=40
t=2, v=40
t=3, v=10  (reset in readback)
t=4, v=30

Without this fix we also see a reset for t=4, v=30 because the chunk is a
non-overlapping in-order chunk and it is used as is.
The fix is to detect when we switch between in-order and out-of-order
chunks and wrap the next chunk in an iterator that clears the
counter reset as if it overlapped with another chunk.

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>