Keep a watermark of Committed series
The current approach is underperforming because all appenders try to
Commit(), regardless they actually created series or not. And even if
they created series, and those series were already committed, it will
still try to grab both locks just to realize there's nothing to commit.

We'll solve that by keeping a watermark of what's already committed, and
keeping track of the series id's created by that appender.
We'll only do the commit operation if there's something to commit.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>