MemPostings.TryCommit()
There's a lot of things to be done in HeadAppender.Commit().
There's no required particular order for many of them.

Sometimes multiple appenders will arrive on the same
MemPostings.Commit() call, both having created series in the past.
Both of them will wait, one of them will just waste that time, because
the first one will most likely commit all the postings needed to be
committed by the second one.

So this adds a MemPostings.TryCommit(), that will do best effort to try
to commit now, and if it's not possible right now (either because
someone is Committing, or someone is Adding), it will retry committing
later, doing other pending work in the meantime (committing samples,
etc.)

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>