Create storage.BatchAppender and return it from DB
`storage.BatchAppender` adds the BatchSeriesRefs method to the Appender().
This method can be used to make sure that all the series from the
incoming scrape or write request already exist (not wired here, as I'm
implementing it for Mimir, but can be used for Prometheus as well).

The reason to add this is to take advantage of a single operation to
reduce locking time on MemPostings among others: right now during Mimir
startup we have to call `Add()` millions of times within a few seconds,
which causes lots of lock contention, but most of the time is actually
spent locking & unlocking rather than adding series.

This method does not implement the optimization yet: it only focuses on
the exposed interface.

I couldn't make Head.Appender() return a BatchAppender because it returns
the initAppender which initializes the head on the first append call:
if I wanted initAppender to implement BatchAppender we would need to add
a timestamp to BatchSeriesRefs, which doesn't really belong here. So we are
exposing HeadInitialized() and InitializeHead() methods and move the
initialization responisibility to the caller, so it can be sure that
BatchAppender() method has always an initialized appender.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>