Implement batches in Appender and Head.getOrCreate
This goes one step further, and does batch series creation in
Appender.BatchSeriesRef as well as creates a new implementation of
getOrCreateBatch().

One of the advantages of the getOrCreateBatch() is that it's now adding
new seires to MemPostings in a batch, probably in a more sorted fashion
than previous implementation.

MemPostings.AddBatch was added but it trivially uses Add() so far.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>