model: add time tracking to Matcher
### Background

In grafana/mimir we sometimes see very expensive regular expressions. Unfortunately, apart from profiling and profiling recording rules, there is no insight into how impactful these regular expressions are.

### What this PR does

Adds `NewMatcherWithTimeTracker` which increments an `*atomic.Duration` when the matcher is executed. This helps to put the time of regular expressions into perspective. For example, we can use it to measure the porion of time of serving a query which went to running regular expressions.

This is an internal version of XXX so that we can move forward with the work while waiting for the upstream PR to be accepted.