labels: Cost & selectivity estimate for Matcher
A very rough sketch of cost estimation for regular expressions. These are being put in a separate file so that they do not conflict with upstream pulls and it's easier to upstream them later.

I've also added tests so that we can track as we make improvements to the model.

The TODOs should be addressed later by XXX

### Regular expressions

By bboreham:

> As part of query planning, estimate the cost of a regexp.
> Very rough and ready, just to get an idea how this might work.
> Add to benchmark so you can compare the estimate against actual
> performance.
>
> Note this is only looking at the Go stdlib implementation; it should
> also make an estimate for those that take the fast path.
>
> Since it looks at `syntax.Regexp` we would either need to do cost
> estimation as part of `newFastRegexMatcher` or retain the pointer so
> we can do it later. Re-parsing every time we want a cost estimate seems
> too expensive.

related to XXX

Co-authored-by: Bryan Boreham <bjboreham@gmail.com>