fix(tsdb): backport zero sample schema optimization from pr 17071 (#1000)
* fix(tsdb): backport zero sample schema optimization from pr 17071

There is an optimization in
https://github.com/prometheus/prometheus/pull/17071
that ensures that the injected zero histogram sample has the same schema
as the next sample.

It turns out that if there are at least three samples:
a normal sample (schema>0), zero sample(schema=0), normal sample (schema>0)
then the histogram rate function will find the lowest schema and normalize
all to that, meaning that the normal samples will be downscaled to schema
0. In dashboards it will look like as if we lost the resolution.

See https://github.com/prometheus/prometheus/blob/9e4d23ddafcdc00021cd8630e78bb819e84ccac9/promql/functions.go#L344

So the optimization is actually needed to not loose resolution on
read back. Alternatively we should use per sample CT as metadata
instead, but that would be a big rewrite of TSDB and PromQL.

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>